# packages/agent-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Appx Agent Server",
    "version": "0.1.0",
    "description": "Pi-SDK-based agent orchestration. Shared auth/model state with explicit, persisted project-scoped session runtimes."
  },
  "components": {
    "schemas": {
...
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "DeploymentTarget": {
        "type": "object",
        "properties": {
          "port": {
            "type": "integer",
            "minimum": 1024,
            "maximum": 65535,
            "description": "Reserved host port (non-privileged, 1024–65535).",
            "example": 10007
          },
          "url": {
            "type": "string",
            "description": "Public URL appx exposes for this environment.",
            "example": "https://eventx.example.com"
          }
        }
      },
      "Deployment": {
        "type": "object",
        "properties": {
          "dev": {
            "$ref": "#/components/schemas/DeploymentTarget"
          },
          "prod": {
            "$ref": "#/components/schemas/DeploymentTarget"
          }
        },
        "description": "Control-plane deployment metadata, if registered."
      },
      "ProjectInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Immutable slug; registry key, route param, and directory name.",
            "example": "my-cool-app"
          },
          "name": {
            "type": "string",
            "example": "My Cool App"
          },
          "projectDir": {
            "type": "string",
            "description": "Absolute working directory under WORKSPACE_DIR.",
            "example": "/workspace/my-cool-app"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO-8601 UTC timestamp",
            "example": "2026-06-03T10:00:00.000Z"
          },
          "deployment": {
            "$ref": "#/components/schemas/Deployment"
          }
        },
        "required": [
          "id",
          "name",
          "projectDir",
          "createdAt"
        ]
      },
      "CreateProjectRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Human-facing project name. Slugified into the immutable id and directory name.",
            "example": "My Cool App"
          },
          "deployment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Deployment"
              },
              {
                "description": "Optional control-plane deployment metadata (dev + prod ports/URLs). Idempotent re-POST updates it."
              }
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "ListProjectsResponse": {
        "type": "object",
        "properties": {
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectInfo"
            }
          }
        },
        "required": [
          "projects"
        ]
      },
      "SessionRow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "01J9Z..."
          },
          "createdAt": {
            "type": "string",
            "description": "ISO-8601 UTC timestamp",
            "example": "2026-05-17T10:00:00.000Z"
          },
          "firstMessage": {
            "type": "string",
            "description": "First user message; empty for never-prompted sessions."
          },
          "messageCount": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "id",
          "createdAt",
          "firstMessage",
          "messageCount"
        ]
      },
      "ListSessionsResponse": {
        "type": "object",
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionRow"
            }
          }
        },
        "required": [
          "sessions"
        ]
      },
      "CreateSessionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/projects": {
      "post": {
        "operationId": "createProject",
        "tags": [
          "projects"
        ],
        "summary": "Create a project, or return the existing one (idempotent on name).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created or already-existing project.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectInfo"
                }
              }
            }
          },
          "400": {
            "description": "Name does not yield a valid project id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listProjects",
        "tags": [
          "projects"
        ],
        "summary": "List registered projects, newest first.",
        "responses": {
          "200": {
            "description": "Registered projects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProjectsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{id}": {
      "get": {
        "operationId": "getProject",
        "tags": [
          "projects"
        ],
        "summary": "Get a single project's metadata.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Project metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectInfo"
                }
              }
            }
          },
          "404": {
            "description": "Unknown project id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteProject",
        "tags": [
          "projects"
        ],
        "summary": "Remove a project: evict runtime, drop metadata, delete working dir + transcripts.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Project removed if it existed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown project id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{projectId}/sessions": {
      "get": {
        "operationId": "listSessions",
        "tags": [
          "sessions"
        ],
        "summary": "List sessions (persisted + in-memory not yet flushed).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "projectId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Sessions, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSessionsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createSession",
        "tags": [
          "sessions"
        ],
        "summary": "Create a new session.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "projectId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Newly created session metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}